+++ Src/Services/Catalog/Catalog.Application/Events/ProductPriceUpdated.cs
@@ -5,6 +5,6 @@ namespace Catalog.Application.Events
     public class ProductPriceUpdated : Event
     {
         public required string ProductId { get; init; }
-        public required decimal NewPrice { get; init; }
+        public required double NewPrice { get; init; }
     }
 }
